+++ examples/insert_and_select.rs
@@ -39,14 +39,14 @@ fn main() {
 fn do_stuff_with_database(db: &mut TransactionMut<MySchema>) {
     // Lets make a new user 'mike',
     let mike = User { name: "mike" };
-    let mike_id = db.insert(mike);
+    let mike_id = db.insert_ok(mike);
 
     // and also insert a dog picture for 'mike'.
     let dog_picture = Image {
         description: "dog",
         uploaded_by: mike_id,
     };
-    let _picture_id = db.insert(dog_picture);
+    let _picture_id = db.insert_ok(dog_picture);
 
     // Now we want to get all pictures for 'mike'.
     let mike_pictures = db.query(|rows| {